Really ugly hack to show a contact's nick name instead of phone number.
If a contact has only a phone number and no name or company name, it seems
that the contact service will set the display name to the first phone
number returned. If this is the case, and the contact has a nickname set
for any account, this patch will display the first one found.

It is ugly and probably should be made less so for real code. It also
should not be considered part of the original Contacts Inspector.
This was just to investigate one way to solve the display problem.